requirements.txt
pymupdf
openai
python-docx